refactor: src 하단에 있는 shop 폴더내 파일들 pages/shop으로 이동
